# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List all photos
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: createPhoto
      tags: [photos]
      summary: Create a photo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/{id}:
    put:
      operationId: updatePhoto
      tags: [photos]
      summary: Update a photo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /instagram/status:
    get:
      operationId: getInstagramStatus
      tags: [instagram]
      summary: Get Instagram connection status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramStatus"
  /instagram/auth:
    get:
      operationId: getInstagramAuthUrl
      tags: [instagram]
      summary: Get Instagram OAuth authorization URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Auth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramAuthUrl"
  /instagram/disconnect:
    post:
      operationId: disconnectInstagram
      tags: [instagram]
      summary: Disconnect Instagram account and clear cached reels
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
  /instagram/sync:
    post:
      operationId: syncInstagramReels
      tags: [instagram]
      summary: Fetch and cache latest reels from Instagram
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sync results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramSyncResponse"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
  /storage/objects/{objectPath}:
    get:
      operationId: getObject
      tags: [storage]
      summary: Serve an object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object file
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Photo:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - url
        - caption
        - category
        - sortOrder
    PhotoInput:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
        category:
          type: string
        sortOrder:
          type: integer
      required:
        - url
...
    InstagramStatus:
      type: object
      properties:
        connected:
          type: boolean
        username:
          type: string
        accountId:
          type: string
        lastSyncAt:
          type: string
          nullable: true
        reelCount:
          type: integer
      required:
        - connected
    InstagramAuthUrl:
      type: object
      properties:
        authUrl:
          type: string
      required:
        - authUrl
...
    InstagramSyncResponse:
      type: object
      properties:
        syncedCount:
          type: integer
        total:
          type: integer
      required:
        - syncedCount
        - total
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath